requirements.txt
langchain-google-genai
langchain-core
langchain-community
duckduckgo-search
feedparser
schedule
requests
python-dotenv
flask
ddgs
yfinance
pandas
numpy